+++ layers-panel/scripts/module.js
@@ -1,0 +1,12 @@
+// ┌───────────────────────────────────┐
+// │  #Events - Global Event Handlers  │
+// ╘═══════════════════════════════════╛
+// When foundry requests a list of controls in top-left
+Hooks.on("getSceneControlButtons", (controls) => {
+    // If active control isn't the drawing menu, do nothing
+    if (ui.controls?.activeControl !== "drawings") { return; }
+    // Remove the 'Clear Drawings' tool
+    const tools = controls.find(control => control.name == ui.controls.activeControl).tools;
+    const pos = tools.findIndex(tool => tool.name == "clear");
+    tools.splice(pos, 1);
+});